release

+++ view/default/plasticzone/src/views/myzone.vue
@@ -128,41 +128,38 @@ export default{
 		},
 		logout: function() {
 			var _this = this;
-			
-			$.ajax({
-				url: version+'/user/logout',
-				type: 'post',
-				data: {
-					token: window.localStorage.getItem("token")
-				},
-				headers: {
-					'X-UA': window.localStorage.getItem("XUA")
-				},
-				dataType: 'JSON'
-			}).then(function(res) {
-				if(res.err == 0) {
-					window.localStorage.setItem("token", "");
-					window.localStorage.setItem("userid", "");
-					weui.alert(res.msg, {
-						title: '塑料圈通讯录',
-						buttons: [{
-							label: '确定',
-							type: 'parimary',
-							onClick: function() {
-								_this.$router.push({
-									name: 'index'
-								});
-							}
-						}]
-					});
-				} else {
-					window.localStorage.setItem("token", "");
-					window.localStorage.setItem("userid", "");
-				}
-				window.localStorage.setItem("XUA","weixin|5.5|"+localStorage.getItem("userid")+"|"+localStorage.getItem("token")+"|0|"+navigator.platform+"|"+navigator.platform+"|"+navigator.platform+"|"+navigator.appName+"|"+navigator.appCodeName+"|0|0|0");
-			}, function() {
+			weui.confirm('确定退出？', function(){ 
+				$.ajax({
+					url: version+'/user/logout',
+					type: 'post',
+					data: {
+						token: window.localStorage.getItem("token")
+					},
+					headers: {
+						'X-UA': window.localStorage.getItem("XUA")
+					},
+					dataType: 'JSON'
+				}).then(function(res) {
+					if(res.err == 0) {
+						window.localStorage.setItem("token", "");
+						window.localStorage.setItem("userid", "");
+						_this.$router.push({
+							name: 'index'
+						});
+					} else {
+						window.localStorage.setItem("token", "");
+						window.localStorage.setItem("userid", "");
+					}
+					window.localStorage.setItem("XUA","weixin|5.5|"+localStorage.getItem("userid")+"|"+localStorage.getItem("token")+"|0|"+navigator.platform+"|"+navigator.platform+"|"+navigator.platform+"|"+navigator.appName+"|"+navigator.appCodeName+"|0|0|0");
+				}, function() {
+	
+				});
+			}, function(){ 
 
-			});
+			}, {
+				title: '塑料圈通讯录'
+			});			
+
 		}
 	},
 	activated: function() {
@@ -227,35 +224,6 @@ export default{
 		}, function() {
 
 		});
-		
-//		$.ajax({
-//			url:version+'/pay/getPrePayOrder',
-//			type:'post',
-//			data:{
-//				type:1,
-//				total_fee:"0.01"
-//			},
-//			headers: {
-//				'X-UA': window.localStorage.getItem("XUA")
-//			},
-//			dataType: 'JSON'
-//		}).done(function(res){
-//			console.log(res);
-//			if(res.err==0){
-//				wx.chooseWXPay({
-//				    timestamp: res.data.timestamp,
-//				    nonceStr: res.data.noncestr,
-//				    package: "prepay_id="+res.data.prepayid,
-//				    signType: 'MD5',
-//				    paySign: res.data.sign,
-//				    success: function (data) {
-//				        console.log(">>>",data);
-//				    }
-//				});			
-//			}
-//		}).fail(function(){
-//			
-//		});
 
 		$.ajax({
 			url: version+'/myInfo/myZone',
